+++ packages/commerce/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @voyant-travel/commerce
 
+## 0.54.1
+
+### Patch Changes
+
+- Updated dependencies [380c46e]
+  - @voyant-travel/bookings@0.248.0
+  - @voyant-travel/catalog@0.259.3
+  - @voyant-travel/distribution@0.229.1
+  - @voyant-travel/finance@0.259.1
+
 ## 0.54.0
 
 ### Minor Changes