+++ packages/cruises-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @voyant-travel/cruises-react
 
+## 0.300.0
+
+### Patch Changes
+
+- @voyant-travel/public-api-react@0.303.0
+- @voyant-travel/catalog-react@0.299.0
+
 ## 0.299.0
 
 ### Patch Changes