+++ packages/finance-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @voyant-travel/finance-react
 
+## 0.301.0
+
+### Patch Changes
+
+- @voyant-travel/bookings-react@0.301.0
+- @voyant-travel/inventory-react@0.183.0
+- @voyant-travel/distribution-react@0.291.0
+- @voyant-travel/operations-react@0.182.0
+
 ## 0.300.0
 
 ### Patch Changes