+++ packages/inventory/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @voyant-travel/inventory
 
+## 0.42.20
+
+### Patch Changes
+
+- Updated dependencies [380c46e]
+  - @voyant-travel/bookings@0.248.0
+  - @voyant-travel/catalog@0.259.3
+  - @voyant-travel/commerce@0.54.1
+  - @voyant-travel/finance@0.259.1
+  - @voyant-travel/operations@0.23.6
+
 ## 0.42.19
 
 ### Patch Changes