+++ packages/relationships-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @voyant-travel/crm-react
 
+## 0.301.0
+
+### Patch Changes
+
+- @voyant-travel/identity-react@0.301.0
+
 ## 0.300.0
 
 ### Patch Changes